Add functionality to restart
When initializing the fluxes, and if the starting time-step is zero, then
we use the initial velocity and initial perturbations (info in the input file
simParams.yaml) to compute the fluxes.
If the starting time-step is not zero, the fluxes are read from the numerical
data saved previously (usinf function `readData` from the namespace `io`).