Do not read pressure and forces anymore
The solution of the pressure and forces (if body) do not have to be
known when re-starting the simulation from a time-step different
from zero. Why? Because they are Lagrange multipliers in the projection
method.
More iterations will be needed for the Poisson solver to converge, but if
it converges in less iterations than the maximum allowed, then we do not
need to read the pressure and the forces.