Merge Gnuplot and Pyplot into one single script
Does not read the time-steps to plot using readSimulationParameters.
The time-steps to plot are either provided by the user in the command-line,
or found listing the solution folders that are in the case directory.

Stores the PNG files generated in the folder `images`, that is created if it
does not already exist.

Adds the flag argument `--gnuplot` to plot the countours of vorticity using
Gnuplot instead of Pyplot.
Merges code from `plotVorticityGnuplot.py`.

Script `plotVorticityGnuplot.py` becomes useless and is deleted.