Change some command-lines arguments
Type flag `--help` to have a list of the new command-line arguments.

This script is ploting the contour of the u- and v- velocities at time-steps
that are provided by the user or by listing the time-step folders that are in
the directory case.

The process is similar to the scripts to plot the contour of vorticity and
contour of pressure.