Remove parameter `restart` from `simParams.yaml`
It the user wants to start the simulation from a tim-step that is not zero,
just set the parameter `startStep` to the value.
Therefore, if `startStep` is not zero, the fluxes will be read from the file,
and the info about iterations and forces will be added to the existing files.