Implement restart functionality
The input parameter `startStep` (file `simParams.yaml`) allows the use to restart
a simulation. When restarting, the fluxes are read from the solution folder that
corresponds to `startStep`. The lambda field, a Lagrange multiplier, is not read.
Finally, the input parameter `nt` (file `simParams.yaml`) is now defined as the number
of time-step to perform, not the final time-step.