Support absolute and relative tolerances for solvers
The user can set the absolute and relative tolerances in the input file
`simParams.yaml` with the keywords `absTolerance` and `relTolerance`,
respectively.
Moreover, the user can set those parameters at execution-time via the
command-line.
For the velocity solver, use `-velocity-atol <value>` and
`-velocity-rtol <value>` to set the absolute and relative tolerances,
respectively.
For the Poisson solver, use `-poisson-atol <value>` and
`-poisson-rtol <value>` to set the absolute and relative tolerances,
respectively.