Add restart functionality and correct nt meaning
You can now restart a simulation by providing the initial time-step
through the parameter `startStep` in the file `simParams.yaml`.

For example, to restart the simulation at save time-step 10,
use the line `startStep: 10` in `simParams.yaml`.

Also, the parameter `nt` is now the number of time-steps to execute;
not the final time-step.