Remove run.info & Use CWD as default simu directory
Do not generate the file `run.info`.
This file only contained information already available in the configuration
files.
This should fix #19.

When the command-line parameter `-caseFolder` is not provided, the current
working directory is used as the simulation directory where all output files
will be saved.

Update installation instructions in README.