Allow to specify Krylov method to use
Fix issue #27

User can now choose the Krylov solver to use for the velocity system and the
Poisson system.
Supported methods are conjugate-gradient (`CG`), bi-conjugate gradient
stabilized (`BICGSTAB`), and GMRES (`GMRES`).
The default solver value is `CG`.

When using `GMRES`, the user has the possibility to choose the restart parameter
in the YAML file `simParams.yaml` with the key `restart` under the node for the
solver.