Make tests to run convection and diffusion tests
`make tests` will run the tests that check the spatial order of convergence
of the convective and diffusive terms.
Possibility to run `make testConvectionTerm` and `make testDiffusionTerm`.

Now each test also write the errors in a temporary file that is read by a
newly-added Python script that computes the observed order of convergence.

Minor change in the input files (added headers).