Downgrade yaml-cpp to 0.5.1
With version 0.5.3, the parse would throw an error when not finding a key in the
configuration file (happening when using `const YAML::Node`).
This behavior was not observed with version 0.5.1.

Example:

```c++

int main(int argc, char **argv)
{
  YAML::Node node = YAML::LoadFile("config.yaml");
  std::cout << "a = " << node["a"].as<float>(1) << std::endl;
  std::cout << "b1 = " << node["b"][0].as<float>(2) << std::endl;
  std::cout << "b2 = " << node["b"][1].as<float>(3) << std::endl;

  const YAML::Node node_const = YAML::LoadFile("config.yaml");
  std::cout << "a = " << node_const["a"].as<float>(1) << std::endl;
  std::cout << "b1 = " << node_const["b"][0].as<float>(2) << std::endl;
  std::cout << "b2 = " << node_const["b"][1].as<float>(3) << std::endl;

  return 0;
} //main
```

YAML-CPP-0.5.1 output:
```
a = 1
b1 = 2
b2 = 3
a = 1
b1 = 2
b2 = 3
```

YAML-CPP-0.5.3 output:
```
a = 1
b1 = 2
b2 = 3
a = 1
terminate called after throwing an instance of 'YAML::InvalidNode'
  what():  invalid node; this may result from using a map iterator as a sequence iterator, or vice-versa
```

Choosing to yaml-cpp-0.5.1 for now.
This should fix #30.